pyproject.toml
[project]
name = "vector-search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastembed>=0.7.4",
    "flagembedding>=1.3.5",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "logging>=0.4.9.6",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "phpserialize3>=0.1.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "requests-cache>=1.2.1",
    "ruff>=0.14.10",
    "scipy-stubs~=1.16.3",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
vector_search = "vector_search.__main__:main"
